+++ pkg/installer/otel_nodejs.go
@@ -8,7 +8,6 @@ import (
 	"os/exec"
 	"path/filepath"
 	"runtime"
-	"sort"
 	"strconv"
 	"strings"
 
@@ -24,288 +23,11 @@ var otelNodePackages = []string{
 	"@opentelemetry/exporter-logs-otlp-http",
 }
 
-func detectNodeProjects() []ScannedProject {
-	projects := scanProjectDirs([]string{"package.json"}, []string{"node_modules"})
-
-	// Expand monorepo workspaces: for each project with a "workspaces" field,
-	// resolve workspace directories and add them as individual projects.
-	var expanded []ScannedProject
-	seen := make(map[string]bool)
-	for _, p := range projects {
-		seen[p.Path] = true
-	}
-	for _, p := range projects {
-		dirs := resolveWorkspaces(p.Path)
-		for _, dir := range dirs {
-			if !seen[dir] {
-				seen[dir] = true
-				expanded = append(expanded, ScannedProject{
-					Path:    dir,
-					Markers: []string{"package.json"},
-				})
-			}
-		}
-	}
-
-	return append(projects, expanded...)
-}
-
-func detectNodeProcesses() []DetectedProcess {
-	return detectProcesses("node", []string{"npm "})
-}
-
-type packageJSON struct {
-	Main            string            `json:"main"`
-	Scripts         map[string]string `json:"scripts"`
-	Dependencies    map[string]string `json:"dependencies"`
-	DevDependencies map[string]string `json:"devDependencies"`
-	Workspaces      json.RawMessage   `json:"workspaces"`
-}
-
-// isNextJSProject checks for Next.js config files or next in package.json dependencies.
-func isNextJSProject(projectPath string) bool {
-	for _, name := range []string{"next.config.js", "next.config.ts", "next.config.mjs"} {
-		if _, err := os.Stat(filepath.Join(projectPath, name)); err == nil {
-			return true
-		}
-	}
-	return hasDependency(projectPath, "next")
-}
-
-// isNuxtProject checks for Nuxt config files or nuxt in package.json dependencies.
-func isNuxtProject(projectPath string) bool {
-	for _, name := range []string{"nuxt.config.js", "nuxt.config.ts", "nuxt.config.mjs", "nuxt.config.mts"} {
-		if _, err := os.Stat(filepath.Join(projectPath, name)); err == nil {
-			return true
-		}
-	}
-	return hasDependency(projectPath, "nuxt")
-}
-
-// hasDependency checks if a package name appears in dependencies or devDependencies.
-func hasDependency(projectPath, pkgName string) bool {
-	data, err := os.ReadFile(filepath.Join(projectPath, "package.json"))
-	if err != nil {
-		return false
-	}
-	var pkg packageJSON
-	if err := json.Unmarshal(data, &pkg); err != nil {
-		return false
-	}
-	if _, ok := pkg.Dependencies[pkgName]; ok {
-		return true
-	}
-	if _, ok := pkg.DevDependencies[pkgName]; ok {
-		return true
-	}
-	return false
-}
-
-// detectNodeFramework returns "next", "nuxt", or "" for a project directory.
-// Next.js takes precedence when both are detected.
-func detectNodeFramework(projectPath string) string {
-	if isNextJSProject(projectPath) {
-		return "next"
-	}
-	if isNuxtProject(projectPath) {
-		return "nuxt"
-	}
-	return ""
-}
-
-// detectNodePackageManager detects the package manager from lockfiles.
-func detectNodePackageManager(projectPath string) string {
-	if _, err := os.Stat(filepath.Join(projectPath, "package-lock.json")); err == nil {
-		logger.Debug("detected package manager", "path", projectPath, "manager", "npm")
-		return "npm"
-	}
-	if _, err := os.Stat(filepath.Join(projectPath, "yarn.lock")); err == nil {
-		logger.Debug("detected package manager", "path", projectPath, "manager", "yarn")
-		return "yarn"
-	}
-	if _, err := os.Stat(filepath.Join(projectPath, "pnpm-lock.yaml")); err == nil {
-		logger.Debug("detected package manager", "path", projectPath, "manager", "pnpm")
-		return "pnpm"
-	}
-	logger.Debug("no lockfile found, defaulting to npm", "path", projectPath)
-	return "npm"
-}
-
-// resolveWorkspaces parses the workspaces field from package.json and returns
-// workspace directories that contain their own package.json.
-func resolveWorkspaces(projectPath string) []string {
-	data, err := os.ReadFile(filepath.Join(projectPath, "package.json"))
-	if err != nil {
-		return nil
-	}
-	var pkg packageJSON
-	if err := json.Unmarshal(data, &pkg); err != nil {
-		return nil
-	}
-	if pkg.Workspaces == nil {
-		return nil
-	}
-
-	var patterns []string
-
-	// Try as array of strings: "workspaces": ["packages/*"]
-	var arr []string
-	if err := json.Unmarshal(pkg.Workspaces, &arr); err == nil {
-		patterns = arr
-	} else {
-		// Try as object: "workspaces": {"packages": ["packages/*"]}
-		var obj struct {
-			Packages []string `json:"packages"`
-		}
-		if err := json.Unmarshal(pkg.Workspaces, &obj); err == nil {
-			patterns = obj.Packages
-		}
-	}
-
-	if len(patterns) == 0 {
-		return nil
-	}
-
-	var workspaceDirs []string
-	for _, pattern := range patterns {
-		matches, err := filepath.Glob(filepath.Join(projectPath, pattern))
-		if err != nil {
-			continue
-		}
-		logger.Debug("resolved workspace pattern", "pattern", pattern, "matchCount", len(matches))
-		for _, match := range matches {
-			info, err := os.Stat(match)
-			if err != nil || !info.IsDir() {
-				continue
-			}
-			if _, err := os.Stat(filepath.Join(match, "package.json")); err == nil {
-				logger.Debug("workspace dir added", "dir", match)
-				workspaceDirs = append(workspaceDirs, match)
-			}
-		}
-	}
-	return workspaceDirs
-}
-
 // generateOtelNodeEnvVars extends base OTel env vars with Node.js-specific settings.
 func generateOtelNodeEnvVars(apiURL, token, serviceName string) map[string]string {
 	envVars := generateBaseOtelEnvVars(apiURL, token, serviceName)
 	envVars["OTEL_NODE_RESOURCE_DETECTORS"] = "all"
 	return envVars
-}
-
-// isJSFileExtension checks if a string ends with a JavaScript/TypeScript file extension.
-func isJSFileExtension(s string) bool {
-	return strings.HasSuffix(s, ".js") || strings.HasSuffix(s, ".ts") ||
-		strings.HasSuffix(s, ".mjs") || strings.HasSuffix(s, ".cjs") ||
-		strings.HasSuffix(s, ".mts") || strings.HasSuffix(s, ".cts")
-}
-
-// extractScriptFile extracts a file reference from a script command string.
-// It looks for tokens ending in .js/.ts/.mjs/.cjs/.mts/.cts that exist on disk.
-func extractScriptFile(projectPath, script string) string {
-	parts := strings.Fields(script)
-	for _, part := range parts {
-		if isJSFileExtension(part) {
-			if _, err := os.Stat(filepath.Join(projectPath, part)); err == nil {
-				return part
-			}
-		}
-	}
-	return ""
-}
-
-// runtimeScriptPrefixes lists package.json script name prefixes that indicate
-// a server/runtime entrypoint (as opposed to build/lint/test tooling). Only
-// scripts matching one of these prefixes are considered when scanning "other
-// scripts" for entrypoints.
-var runtimeScriptPrefixes = []string{"start", "dev", "serve", "server"}
-
-// isRuntimeScript checks if a package.json script name looks like a runtime
-// entrypoint (e.g. "start:api", "dev", "serve:frontend") rather than a
-// build/lint/test script.
-func isRuntimeScript(name string) bool {
-	lower := strings.ToLower(name)
-	for _, prefix := range runtimeScriptPrefixes {
-		if lower == prefix || strings.HasPrefix(lower, prefix+":") {
-			return true
-		}
-	}
-	return false
-}
-
-func detectNodeEntrypoints(projectPath string) []string {
-	// For framework projects, return marker entrypoints.
-	framework := detectNodeFramework(projectPath)
-	if framework == "next" {
-		logger.Debug("node entrypoint: Next.js project detected", "path", projectPath)
-		return []string{"next:start"}
-	}
-	if framework == "nuxt" {
-		logger.Debug("node entrypoint: Nuxt project detected", "path", projectPath)
-		return []string{"nuxt:start"}
-	}
-
-	data, err := os.ReadFile(filepath.Join(projectPath, "package.json"))
-	if err != nil {
-		return nil
-	}
-
-	var pkg packageJSON
-	_ = json.Unmarshal(data, &pkg)
-
-	if pkg.Main != "" {
-		logger.Debug("node entrypoint: checking 'main' field", "main", pkg.Main)
-		if _, err := os.Stat(filepath.Join(projectPath, pkg.Main)); err == nil {
-			logger.Debug("node entrypoint found via 'main'", "file", pkg.Main)
-			return []string{pkg.Main}
-		}
-	}
-
-	if start, ok := pkg.Scripts["start"]; ok && start != "" {
-		logger.Debug("node entrypoint: checking 'scripts.start'", "start", start)
-		if found := extractScriptFile(projectPath, start); found != "" {
-			logger.Debug("node entrypoint found via 'scripts.start'", "file", found)
-			return []string{found}
-		}
-	}
-
-	// Scan runtime-like scripts (start:*, dev:*, serve:*, server:*) for file
-	// references. Non-runtime scripts (build, lint, test, etc.) are skipped to
-	// avoid picking up config files as entrypoints.
-	if len(pkg.Scripts) > 0 {
-		seen := make(map[string]bool)
-		var otherEntrypoints []string
-		for name, script := range pkg.Scripts {
-			if name == "start" || script == "" {
-				continue
-			}
-			if !isRuntimeScript(name) {
-				continue
-			}
-			if found := extractScriptFile(projectPath, script); found != "" && !seen[found] {
-				seen[found] = true
-				otherEntrypoints = append(otherEntrypoints, found)
-				logger.Debug("node entrypoint found via script", "script", name, "file", found)
-			}
-		}
-		if len(otherEntrypoints) > 0 {
-			sort.Strings(otherEntrypoints)
-			return otherEntrypoints
-		}
-	}
-
-	for _, base := range []string{"index", "app", "server"} {
-		for _, ext := range []string{".js", ".ts", ".mjs", ".cjs", ".mts", ".cts"} {
-			name := base + ext
-			if _, err := os.Stat(filepath.Join(projectPath, name)); err == nil {
-				logger.Debug("node entrypoint found via fallback", "file", name)
-				return []string{name}
-			}
-		}
-	}
-
-	return nil
 }
 
 type NodeInstrumentationPlan struct {
@@ -658,7 +380,6 @@ func (p *NodeInstrumentationPlan) Execute() {
 	fmt.Println("  Waiting for traffic — send requests to your services to generate traces and metrics.")
 }
 
-// launchEntrypoint
 // launchEntrypoint starts a managed process for a single entrypoint.
 func launchEntrypoint(svcName, projectPath, entrypoint string, cmd *exec.Cmd) *ManagedProcess {
 	logName := svcName + ".log"
