+++ pkg/installer/otel_nodejs_project.go
@@ -1,0 +1,172 @@
+package installer
+
+import (
+	"encoding/json"
+	"os"
+	"path/filepath"
+
+	"github.com/dynatrace-oss/dtwiz/pkg/logger"
+)
+
+type packageJSON struct {
+	Main            string            `json:"main"`
+	Scripts         map[string]string `json:"scripts"`
+	Dependencies    map[string]string `json:"dependencies"`
+	DevDependencies map[string]string `json:"devDependencies"`
+	Workspaces      json.RawMessage   `json:"workspaces"`
+}
+
+func detectNodeProjects() []ScannedProject {
+	projects := scanProjectDirs([]string{"package.json"}, []string{"node_modules"})
+
+	// Expand monorepo workspaces: for each project with a "workspaces" field,
+	// resolve workspace directories and add them as individual projects.
+	var expanded []ScannedProject
+	seen := make(map[string]bool)
+	for _, p := range projects {
+		seen[p.Path] = true
+	}
+	for _, p := range projects {
+		dirs := resolveWorkspaces(p.Path)
+		for _, dir := range dirs {
+			if !seen[dir] {
+				seen[dir] = true
+				expanded = append(expanded, ScannedProject{
+					Path:    dir,
+					Markers: []string{"package.json"},
+				})
+			}
+		}
+	}
+
+	return append(projects, expanded...)
+}
+
+func detectNodeProcesses() []DetectedProcess {
+	return detectProcesses("node", []string{"npm "})
+}
+
+// isNextJSProject checks for Next.js config files or next in package.json dependencies.
+func isNextJSProject(projectPath string) bool {
+	for _, name := range []string{"next.config.js", "next.config.ts", "next.config.mjs"} {
+		if _, err := os.Stat(filepath.Join(projectPath, name)); err == nil {
+			return true
+		}
+	}
+	return hasDependency(projectPath, "next")
+}
+
+// isNuxtProject checks for Nuxt config files or nuxt in package.json dependencies.
+func isNuxtProject(projectPath string) bool {
+	for _, name := range []string{"nuxt.config.js", "nuxt.config.ts", "nuxt.config.mjs", "nuxt.config.mts"} {
+		if _, err := os.Stat(filepath.Join(projectPath, name)); err == nil {
+			return true
+		}
+	}
+	return hasDependency(projectPath, "nuxt")
+}
+
+// hasDependency checks if a package name appears in dependencies or devDependencies.
+func hasDependency(projectPath, pkgName string) bool {
+	data, err := os.ReadFile(filepath.Join(projectPath, "package.json"))
+	if err != nil {
+		return false
+	}
+	var pkg packageJSON
+	if err := json.Unmarshal(data, &pkg); err != nil {
+		return false
+	}
+	if _, ok := pkg.Dependencies[pkgName]; ok {
+		return true
+	}
+	if _, ok := pkg.DevDependencies[pkgName]; ok {
+		return true
+	}
+	return false
+}
+
+// detectNodeFramework returns "next", "nuxt", or "" for a project directory.
+// Next.js takes precedence when both are detected.
+func detectNodeFramework(projectPath string) string {
+	if isNextJSProject(projectPath) {
+		return "next"
+	}
+	if isNuxtProject(projectPath) {
+		return "nuxt"
+	}
+	return ""
+}
+
+// detectNodePackageManager detects the package manager from lockfiles.
+func detectNodePackageManager(projectPath string) string {
+	if _, err := os.Stat(filepath.Join(projectPath, "package-lock.json")); err == nil {
+		logger.Debug("detected package manager", "path", projectPath, "manager", "npm")
+		return "npm"
+	}
+	if _, err := os.Stat(filepath.Join(projectPath, "yarn.lock")); err == nil {
+		logger.Debug("detected package manager", "path", projectPath, "manager", "yarn")
+		return "yarn"
+	}
+	if _, err := os.Stat(filepath.Join(projectPath, "pnpm-lock.yaml")); err == nil {
+		logger.Debug("detected package manager", "path", projectPath, "manager", "pnpm")
+		return "pnpm"
+	}
+	logger.Debug("no lockfile found, defaulting to npm", "path", projectPath)
+	return "npm"
+}
+
+// resolveWorkspaces parses the workspaces field from package.json and returns
+// workspace directories that contain their own package.json.
+func resolveWorkspaces(projectPath string) []string {
+	data, err := os.ReadFile(filepath.Join(projectPath, "package.json"))
+	if err != nil {
+		return nil
+	}
+	var pkg packageJSON
+	if err := json.Unmarshal(data, &pkg); err != nil {
+		return nil
+	}
+	if pkg.Workspaces == nil {
+		return nil
+	}
+
+	var patterns []string
+
+	// Try as array of strings: "workspaces": ["packages/*"]
+	var arr []string
+	if err := json.Unmarshal(pkg.Workspaces, &arr); err == nil {
+		patterns = arr
+	} else {
+		// Try as object: "workspaces": {"packages": ["packages/*"]}
+		var obj struct {
+			Packages []string `json:"packages"`
+		}
+		if err := json.Unmarshal(pkg.Workspaces, &obj); err == nil {
+			patterns = obj.Packages
+		}
+	}
+
+	if len(patterns) == 0 {
+		return nil
+	}
+
+	var workspaceDirs []string
+	for _, pattern := range patterns {
+		matches, err := filepath.Glob(filepath.Join(projectPath, pattern))
+		if err != nil {
+			continue
+		}
+		logger.Debug("resolved workspace pattern", "pattern", pattern, "matchCount", len(matches))
+		for _, match := range matches {
+			info, err := os.Stat(match)
+			if err != nil || !info.IsDir() {
+				continue
+			}
+			if _, err := os.Stat(filepath.Join(match, "package.json")); err == nil {
+				logger.Debug("workspace dir added", "dir", match)
+				workspaceDirs = append(workspaceDirs, match)
+			}
+		}
+	}
+	return workspaceDirs
+}
