+++ pkg/installer/otel_nodejs_test.go
@@ -9,123 +9,6 @@ import (
 	"testing"
 )
 
-func TestDetectNodeProjects_Found(t *testing.T) {
-	dir := t.TempDir()
-	realDir, _ := filepath.EvalSymlinks(dir)
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{"name":"app"}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	setTestWorkingDir(t, dir)
-	projects := detectNodeProjects()
-	found := false
-	for _, p := range projects {
-		if p.Path == dir || p.Path == realDir {
-			found = true
-		}
-	}
-	if !found {
-		t.Errorf("expected project at %s, got %v", dir, projects)
-	}
-}
-
-func TestDetectNodeProjects_ExcludesNodeModules(t *testing.T) {
-	dir := t.TempDir()
-	// Create node_modules subdirectory with a package.json inside.
-	nmDir := filepath.Join(dir, "node_modules", "somelib")
-	if err := os.MkdirAll(nmDir, 0755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(nmDir, "package.json"), []byte(`{}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-	// Create the real project package.json.
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{"name":"app"}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	setTestWorkingDir(t, dir)
-	projects := detectNodeProjects()
-	for _, p := range projects {
-		if filepath.Base(filepath.Dir(p.Path)) == "node_modules" {
-			t.Errorf("node_modules project should be excluded, found: %s", p.Path)
-		}
-	}
-}
-
-func TestDetectNodeEntrypoints_Main(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{"main":"server.js"}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(dir, "server.js"), []byte(""), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	eps := detectNodeEntrypoints(dir)
-	if len(eps) == 0 || eps[0] != "server.js" {
-		t.Errorf("expected [server.js], got %v", eps)
-	}
-}
-
-func TestDetectNodeEntrypoints_ScriptsStart(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{"scripts":{"start":"node app.js"}}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(dir, "app.js"), []byte(""), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	eps := detectNodeEntrypoints(dir)
-	if len(eps) == 0 || eps[0] != "app.js" {
-		t.Errorf("expected [app.js], got %v", eps)
-	}
-}
-
-func TestDetectNodeEntrypoints_Fallback(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(dir, "index.js"), []byte(""), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	eps := detectNodeEntrypoints(dir)
-	if len(eps) == 0 || eps[0] != "index.js" {
-		t.Errorf("expected [index.js], got %v", eps)
-	}
-}
-
-func TestDetectNodeEntrypoints_TypeScript(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-	// Only a TypeScript variant exists.
-	if err := os.WriteFile(filepath.Join(dir, "app.ts"), []byte(""), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	eps := detectNodeEntrypoints(dir)
-	if len(eps) == 0 || eps[0] != "app.ts" {
-		t.Errorf("expected [app.ts], got %v", eps)
-	}
-}
-
-func TestDetectNodeEntrypoints_None(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	eps := detectNodeEntrypoints(dir)
-	if len(eps) != 0 {
-		t.Errorf("expected empty entrypoints, got %v", eps)
-	}
-}
-
 func TestBuildNodeInstrumentationPlan(t *testing.T) {
 	apiURL := "https://tenant.live.dynatrace.com"
 	token := "token"
@@ -460,104 +343,6 @@ func TestNodeInstrumentationPlan_PrintPlanSteps_ShowsRunningPIDs(t *testing.T) {
 	}
 }
 
-// --- Task 1.1a: isNextJSProject tests ---
-
-func TestIsNextJSProject_ConfigFile(t *testing.T) {
-	for _, configFile := range []string{"next.config.js", "next.config.ts", "next.config.mjs"} {
-		t.Run(configFile, func(t *testing.T) {
-			dir := t.TempDir()
-			if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{"name":"app"}`), 0644); err != nil {
-				t.Fatal(err)
-			}
-			if err := os.WriteFile(filepath.Join(dir, configFile), []byte(""), 0644); err != nil {
-				t.Fatal(err)
-			}
-			if !isNextJSProject(dir) {
-				t.Errorf("expected isNextJSProject=true for %s", configFile)
-			}
-		})
-	}
-}
-
-func TestIsNextJSProject_PackageDep(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{"dependencies":{"next":"14.0.0"}}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-	if !isNextJSProject(dir) {
-		t.Error("expected isNextJSProject=true for next in dependencies")
-	}
-}
-
-func TestIsNextJSProject_DevDep(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{"devDependencies":{"next":"14.0.0"}}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-	if !isNextJSProject(dir) {
-		t.Error("expected isNextJSProject=true for next in devDependencies")
-	}
-}
-
-func TestIsNextJSProject_NotNextJS(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{"dependencies":{"express":"4.0.0"}}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-	if isNextJSProject(dir) {
-		t.Error("expected isNextJSProject=false for non-Next.js project")
-	}
-}
-
-// --- Task 1.1b: detectNodeFramework tests ---
-
-func TestDetectNodeFramework_NuxtConfigFile(t *testing.T) {
-	for _, configFile := range []string{"nuxt.config.js", "nuxt.config.ts", "nuxt.config.mjs"} {
-		t.Run(configFile, func(t *testing.T) {
-			dir := t.TempDir()
-			if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{"name":"app"}`), 0644); err != nil {
-				t.Fatal(err)
-			}
-			if err := os.WriteFile(filepath.Join(dir, configFile), []byte(""), 0644); err != nil {
-				t.Fatal(err)
-			}
-			if got := detectNodeFramework(dir); got != "nuxt" {
-				t.Errorf("detectNodeFramework() = %q, want %q for %s", got, "nuxt", configFile)
-			}
-		})
-	}
-}
-
-func TestDetectNodeFramework_NuxtDep(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{"dependencies":{"nuxt":"3.0.0"}}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-	if got := detectNodeFramework(dir); got != "nuxt" {
-		t.Errorf("detectNodeFramework() = %q, want %q", got, "nuxt")
-	}
-}
-
-func TestDetectNodeFramework_NextTakesPrecedence(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{"dependencies":{"next":"14.0.0","nuxt":"3.0.0"}}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-	if got := detectNodeFramework(dir); got != "next" {
-		t.Errorf("detectNodeFramework() = %q, want %q (Next.js takes precedence)", got, "next")
-	}
-}
-
-func TestDetectNodeFramework_Regular(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{"dependencies":{"express":"4.0.0"}}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-	if got := detectNodeFramework(dir); got != "" {
-		t.Errorf("detectNodeFramework() = %q, want empty string for regular project", got)
-	}
-}
-
 // --- Prerequisite check: node_modules/ ---
 
 func TestExecute_RegularApp_MissingNodeModules_DoesNotCreateOtelDir(t *testing.T) {
@@ -608,345 +393,6 @@ func TestExecute_NextJSApp_MissingNodeModules_DoesNotCreateOtelDir(t *testing.T)
 
 	if _, err := os.Stat(otelDir); !os.IsNotExist(err) {
 		t.Errorf(".otel/ was created despite missing node_modules/ — expected early exit")
-	}
-}
-
-// --- Task 1.2: detectNodePackageManager tests ---
-
-func TestDetectNodePackageManager_NPM(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "package-lock.json"), []byte("{}"), 0644); err != nil {
-		t.Fatal(err)
-	}
-	if got := detectNodePackageManager(dir); got != "npm" {
-		t.Errorf("detectNodePackageManager() = %q, want %q", got, "npm")
-	}
-}
-
-func TestDetectNodePackageManager_Yarn(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "yarn.lock"), []byte(""), 0644); err != nil {
-		t.Fatal(err)
-	}
-	if got := detectNodePackageManager(dir); got != "yarn" {
-		t.Errorf("detectNodePackageManager() = %q, want %q", got, "yarn")
-	}
-}
-
-func TestDetectNodePackageManager_PNPM(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "pnpm-lock.yaml"), []byte(""), 0644); err != nil {
-		t.Fatal(err)
-	}
-	if got := detectNodePackageManager(dir); got != "pnpm" {
-		t.Errorf("detectNodePackageManager() = %q, want %q", got, "pnpm")
-	}
-}
-
-func TestDetectNodePackageManager_Default(t *testing.T) {
-	dir := t.TempDir()
-	if got := detectNodePackageManager(dir); got != "npm" {
-		t.Errorf("detectNodePackageManager() = %q, want %q (default)", got, "npm")
-	}
-}
-
-// --- Task 1.3: Monorepo detection tests ---
-
-func TestDetectNodeProjects_Monorepo(t *testing.T) {
-	dir := t.TempDir()
-	realDir, _ := filepath.EvalSymlinks(dir)
-
-	// Root package.json with workspaces.
-	rootPkg := `{"name":"monorepo","workspaces":["packages/*"]}`
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(rootPkg), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	// Create workspace packages.
-	for _, pkg := range []string{"api", "web"} {
-		pkgDir := filepath.Join(dir, "packages", pkg)
-		if err := os.MkdirAll(pkgDir, 0755); err != nil {
-			t.Fatal(err)
-		}
-		if err := os.WriteFile(filepath.Join(pkgDir, "package.json"), []byte(`{"name":"`+pkg+`"}`), 0644); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	setTestWorkingDir(t, dir)
-	projects := detectNodeProjects()
-
-	// Should include the root and both workspace packages.
-	paths := make(map[string]bool)
-	for _, p := range projects {
-		paths[p.Path] = true
-	}
-
-	// Check root is present.
-	if !paths[dir] && !paths[realDir] {
-		t.Errorf("expected monorepo root in projects, got %v", projects)
-	}
-
-	// Check workspace packages are present.
-	for _, pkg := range []string{"api", "web"} {
-		pkgPath := filepath.Join(dir, "packages", pkg)
-		realPkgPath := filepath.Join(realDir, "packages", pkg)
-		if !paths[pkgPath] && !paths[realPkgPath] {
-			t.Errorf("expected workspace package %s in projects, got %v", pkg, projects)
-		}
-	}
-}
-
-func TestResolveWorkspaces_ArrayFormat(t *testing.T) {
-	dir := t.TempDir()
-
-	rootPkg := `{"workspaces":["packages/*"]}`
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(rootPkg), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	pkgDir := filepath.Join(dir, "packages", "lib")
-	if err := os.MkdirAll(pkgDir, 0755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(pkgDir, "package.json"), []byte(`{}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	dirs := resolveWorkspaces(dir)
-	if len(dirs) != 1 {
-		t.Fatalf("expected 1 workspace dir, got %v", dirs)
-	}
-	if dirs[0] != pkgDir {
-		t.Errorf("expected %q, got %q", pkgDir, dirs[0])
-	}
-}
-
-func TestResolveWorkspaces_ObjectFormat(t *testing.T) {
-	dir := t.TempDir()
-
-	rootPkg := `{"workspaces":{"packages":["packages/*"]}}`
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(rootPkg), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	pkgDir := filepath.Join(dir, "packages", "lib")
-	if err := os.MkdirAll(pkgDir, 0755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(pkgDir, "package.json"), []byte(`{}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	dirs := resolveWorkspaces(dir)
-	if len(dirs) != 1 {
-		t.Fatalf("expected 1 workspace dir, got %v", dirs)
-	}
-	if dirs[0] != pkgDir {
-		t.Errorf("expected %q, got %q", pkgDir, dirs[0])
-	}
-}
-
-func TestResolveWorkspaces_NoWorkspaces(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{"name":"app"}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	dirs := resolveWorkspaces(dir)
-	if len(dirs) != 0 {
-		t.Errorf("expected no workspaces, got %v", dirs)
-	}
-}
-
-func TestResolveWorkspaces_SkipsDirWithoutPackageJSON(t *testing.T) {
-	dir := t.TempDir()
-
-	rootPkg := `{"workspaces":["packages/*"]}`
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(rootPkg), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	// Create a workspace dir without package.json.
-	emptyDir := filepath.Join(dir, "packages", "empty")
-	if err := os.MkdirAll(emptyDir, 0755); err != nil {
-		t.Fatal(err)
-	}
-
-	dirs := resolveWorkspaces(dir)
-	if len(dirs) != 0 {
-		t.Errorf("expected no workspaces (no package.json), got %v", dirs)
-	}
-}
-
-// --- Task 1.4/1.5: detectNodeEntrypoints for Next.js / Nuxt ---
-
-func TestDetectNodeEntrypoints_NextJS(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{"dependencies":{"next":"14.0.0"}}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	eps := detectNodeEntrypoints(dir)
-	if len(eps) != 1 || eps[0] != "next:start" {
-		t.Errorf("expected [next:start], got %v", eps)
-	}
-}
-
-func TestDetectNodeEntrypoints_Nuxt(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(`{"dependencies":{"nuxt":"3.0.0"}}`), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	eps := detectNodeEntrypoints(dir)
-	if len(eps) != 1 || eps[0] != "nuxt:start" {
-		t.Errorf("expected [nuxt:start], got %v", eps)
-	}
-}
-
-// --- Other scripts entrypoint detection ---
-
-func TestDetectNodeEntrypoints_OtherScripts(t *testing.T) {
-	dir := t.TempDir()
-	// scripts.start uses npm-run-all (no direct file reference),
-	// but sub-scripts reference actual files.
-	pkgJSON := `{
-		"scripts": {
-			"start": "npm-run-all --parallel start:frontend start:order start:delivery",
-			"start:frontend": "node s-frontend/app.js",
-			"start:order": "node s-order/app.js",
-			"start:delivery": "node s-delivery/app.js"
-		}
-	}`
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(pkgJSON), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	// Create the entrypoint files.
-	for _, sub := range []string{"s-frontend", "s-order", "s-delivery"} {
-		subDir := filepath.Join(dir, sub)
-		if err := os.MkdirAll(subDir, 0755); err != nil {
-			t.Fatal(err)
-		}
-		if err := os.WriteFile(filepath.Join(subDir, "app.js"), []byte(""), 0644); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	eps := detectNodeEntrypoints(dir)
-	if len(eps) != 3 {
-		t.Fatalf("expected 3 entrypoints, got %v", eps)
-	}
-
-	// Should be sorted.
-	want := []string{"s-delivery/app.js", "s-frontend/app.js", "s-order/app.js"}
-	for i, ep := range eps {
-		if ep != want[i] {
-			t.Errorf("entrypoints[%d] = %q, want %q", i, ep, want[i])
-		}
-	}
-}
-
-func TestDetectNodeEntrypoints_OtherScripts_SkipsMissing(t *testing.T) {
-	dir := t.TempDir()
-	// Script references a file that doesn't exist on disk.
-	pkgJSON := `{
-		"scripts": {
-			"start": "npm-run-all --parallel start:api",
-			"start:api": "node src/api.js"
-		}
-	}`
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(pkgJSON), 0644); err != nil {
-		t.Fatal(err)
-	}
-
-	eps := detectNodeEntrypoints(dir)
-	if len(eps) != 0 {
-		t.Errorf("expected 0 entrypoints (file missing), got %v", eps)
-	}
-}
-
-func TestDetectNodeEntrypoints_IgnoresNonRuntimeScripts(t *testing.T) {
-	dir := t.TempDir()
-	// Only non-runtime scripts reference existing files.
-	// These should NOT be treated as entrypoints.
-	pkgJSON := `{
-		"scripts": {
-			"lint": "eslint --config eslint.config.js src/",
-			"build": "esbuild build.config.ts",
-			"test": "jest jest.config.js",
-			"typecheck": "tsc --project tsconfig.json"
-		}
-	}`
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(pkgJSON), 0644); err != nil {
-		t.Fatal(err)
-	}
-	for _, f := range []string{"eslint.config.js", "build.config.ts", "jest.config.js", "tsconfig.json"} {
-		if err := os.WriteFile(filepath.Join(dir, f), []byte(""), 0644); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	eps := detectNodeEntrypoints(dir)
-	if len(eps) != 0 {
-		t.Errorf("expected 0 entrypoints (non-runtime scripts), got %v", eps)
-	}
-}
-
-func TestDetectNodeEntrypoints_RuntimeScriptPrefixes(t *testing.T) {
-	dir := t.TempDir()
-	// All of these script names should be scanned for entrypoints.
-	pkgJSON := `{
-		"scripts": {
-			"dev": "node dev-server.js",
-			"serve:api": "node api.js",
-			"server:worker": "node worker.js"
-		}
-	}`
-	if err := os.WriteFile(filepath.Join(dir, "package.json"), []byte(pkgJSON), 0644); err != nil {
-		t.Fatal(err)
-	}
-	for _, f := range []string{"dev-server.js", "api.js", "worker.js"} {
-		if err := os.WriteFile(filepath.Join(dir, f), []byte(""), 0644); err != nil {
-			t.Fatal(err)
-		}
-	}
-
-	eps := detectNodeEntrypoints(dir)
-	if len(eps) != 3 {
-		t.Fatalf("expected 3 entrypoints, got %v", eps)
-	}
-}
-
-func TestIsRuntimeScript(t *testing.T) {
-	cases := []struct {
-		name string
-		want bool
-	}{
-		{"start", true},
-		{"start:api", true},
-		{"start:frontend", true},
-		{"dev", true},
-		{"dev:watch", true},
-		{"serve", true},
-		{"serve:api", true},
-		{"server", true},
-		{"server:worker", true},
-		{"build", false},
-		{"test", false},
-		{"lint", false},
-		{"typecheck", false},
-		{"format", false},
-		{"prestart", false},
-		{"postinstall", false},
-		{"starting", false},
-		{"developer", false},
-	}
-	for _, tc := range cases {
-		if got := isRuntimeScript(tc.name); got != tc.want {
-			t.Errorf("isRuntimeScript(%q) = %v, want %v", tc.name, got, tc.want)
-		}
 	}
 }
